[1.4.x] Fixed #24238 -- Removed unused WSGIRequestHandler.get_environ()
Also moved the test as it wasn't running.
